prepare release

+++ src/Service/ThemesService.php
@@ -148,29 +148,21 @@ class ThemesService implements ThemesServiceInterface
         }
         $name = $postData['file']->getClientFileName();
         $postData['file']->moveTo(TMP . $name);
-        $srcDirName = basename($name, '.zip');
+        $srcName = basename($name, '.zip');
         $zip = new BcZip();
         if (!$zip->extract(TMP . $name, TMP)) {
             throw new BcException(__d('baser_core', 'アップロードしたZIPファイルの展開に失敗しました。'));
         }
 
-        $dstName = $srcName = Inflector::camelize($srcDirName);
-        if (preg_match('/^(.+?)([0-9]+)$/', $srcName, $matches)) {
-            $baseName = $matches[1];
-            $num = $matches[2];
-        } else {
-            $baseName = $srcName;
-            $num = null;
-        }
+        $num = 2;
+        $dstName = Inflector::camelize($srcName);
         while(is_dir(BASER_THEMES . $dstName) || is_dir(BASER_THEMES . Inflector::dasherize($dstName))) {
-            if (is_null($num)) $num = 1;
+            $dstName = Inflector::camelize($srcName) . $num;
             $num++;
-            $dstName = $baseName . $num;
         }
-        $folder = new Folder(TMP . $srcDirName);
+        $folder = new Folder(TMP . $srcName);
         $folder->move(BASER_THEMES . $dstName, ['mode' => 0777]);
         unlink(TMP . $name);
-        BcUtil::changePluginClassName($srcName, $dstName);
         BcUtil::changePluginNameSpace($dstName);
         return $dstName;
     }
@@ -367,8 +359,7 @@ class ThemesService implements ThemesServiceInterface
         if (!is_writable(BASER_THEMES)) throw new BcException(BASER_THEMES . ' に書込み権限がありません。');
 
         if (in_array($theme, Configure::read('BcApp.core'))) $theme = Inflector::dasherize($theme);
-        $oldTheme = Inflector::camelize($theme, '-');
-        $newTheme = $oldTheme . 'Copy';
+        $newTheme = Inflector::camelize($theme, '-') . 'Copy';
         while(true) {
             if (!is_dir(BASER_THEMES . $newTheme)) break;
             $newTheme .= 'Copy';
@@ -379,21 +370,7 @@ class ThemesService implements ThemesServiceInterface
         ])) {
             return false;
         }
-        if(!file_exists(BASER_THEMES . $newTheme . DS . 'src' . DS . $newTheme . 'Plugin.php')) {
-            if (file_exists(BASER_THEMES . $newTheme . DS . 'src' . DS . 'Plugin.php')) {
-                rename(
-                    BASER_THEMES . $newTheme . DS . 'src' . DS . 'Plugin.php',
-                    BASER_THEMES . $newTheme . DS . 'src' . DS . $newTheme . 'Plugin.php'
-                );
-            } elseif (file_exists(BASER_THEMES . $newTheme . DS . 'src' . DS . $oldTheme . 'Plugin.php')) {
-                rename(
-                    BASER_THEMES . $newTheme . DS . 'src' . DS . $oldTheme . 'Plugin.php',
-                    BASER_THEMES . $newTheme . DS . 'src' . DS . $newTheme . 'Plugin.php'
-                );
-            }
-        }
-        BcUtil::changePluginClassName($oldTheme, $newTheme);
-        BcUtil::changePluginNameSpace($newTheme);
+        if(!BcUtil::changePluginNameSpace($newTheme)) return false;
         return true;
     }
 
@@ -479,14 +456,7 @@ class ThemesService implements ThemesServiceInterface
             $this->_writeCsv($plugin, $tmpDir . $plugin . DS, $excludes);
             $folder = new Folder($tmpDir . $plugin);
             $files = $folder->read();
-            if (!$files[0] && !$files[1]) {
-                $folder->delete($tmpDir . $plugin);
-            } else {
-                $pluginClass = Plugin::getCollection()->get($plugin);
-                if(method_exists($pluginClass, 'modifyDownloadDefaultData')) {
-                    $pluginClass->modifyDownloadDefaultData($tmpDir . $plugin . DS);
-                }
-            }
+            if (!$files[0] && !$files[1]) $folder->delete($tmpDir . $plugin);
         }
         // site_configs 調整
         $this->_modifySiteConfigsCsv($tmpDir . 'BaserCore' . DS . 'site_configs.csv');
